ENH: Provide consistent floating point comparisons
A set of floating point comparison tools is provided
so that we can uniformly implement the various
comparisons of floating point numbers.

In some cases we desire bit-wise equality for
floating point numbers.  In other cases we
desire ApproximatelyEqual.

In particular, when comparing the equality of
a floating point number to an integer, it is
often desired to check within a small tolerance
of equality.

Change-Id: I376372edcb766e426a2ed1cde86856d890ed8650